Attachments: refactor builder specs
